// src/lib/api/handler/error-codes.ts
// ErrorCodes.ts
enum ErrorCode {
    SUCCESS = 0,
  
    // Client Errors (1xx)
    BAD_REQUEST = 100,
    UNAUTHORIZED = 101,
    ALREADY_VERIFIED = 102,
    TOO_MANY_REQUEST = 103,
  
    // Authentication Errors (2xx)
    INVALID_EMAIL_OR_PASSWORD = 200,
    EMAIL_NOT_VERIFIED = 201,
    INVALID_PASSWORD = 202,
    PASSWORD_RESET_CODE_EXPIRED = 203,
    INVALID_RESET_CODE = 204,
    BCRYPT_CANT_HASH = 205,
    TOKEN_GENERATION_ERROR = 206,
  
    // User-Related Errors (3xx)
    USER_NOT_FOUND = 300,
    CANNOT_CREATE_USER = 301,
    CANNOT_UPDATE_USER = 302,
    CANNOT_DELETE_USER = 303,
    CANNOT_FETCH_USER = 304,
    CANNOT_UPDATE_USER_IMAGE = 305,
    CANNOT_GET_USER_IMAGE_NAME = 306,
    CANNOT_DELETE_IMAGE = 307,
    UUID_CONVERSION_ERR = 308,
  
    // Employee-Related Errors (4xx)
    EMPLOYEE_NOT_FOUND = 400,
    EMPLOYEE_REPO_ERR = 401,
    CANNOT_CREATE_EMPLOYEE = 402,
    CANNOT_UPDATE_EMPLOYEE = 403,
    CANNOT_DELETE_EMPLOYEE = 404,
    CANNOT_FETCH_EMPLOYEE = 405,
  
    // Manager-Related Errors (5xx)
    MANAGER_NOT_FOUND = 500,
    CANNOT_CREATE_MANAGER = 501,
    CANNOT_UPDATE_MANAGER = 502,
    CANNOT_DELETE_MANAGER = 503,
    CANNOT_FETCH_MANAGER = 504,
  
    // Company-Related Errors (6xx)
    COMPANY_NOT_FOUND = 600,
    COMPANY_REPO_ERR = 601,
    CANNOT_CREATE_COMPANY = 602,
    CANNOT_UPDATE_COMPANY = 603,
    CANNOT_DELETE_COMPANY = 604,
    CANNOT_FETCH_COMPANY = 605,
    CANNOT_FETCH_COMPANY_IMAGE = 606,
    CANNOT_UPDATE_COMPANY_IMAGE = 607,
  
    // Workplace-Related Errors (7xx)
    WORKPLACE_NOT_FOUND = 700,
    CANNOT_CREATE_WORKPLACE = 701,
    CANNOT_UPDATE_WORKPLACE = 702,
    CANNOT_DELETE_WORKPLACE = 703,
    CANNOT_FETCH_WORKPLACE = 704,
  
    // Role-Related Errors (8xx)
    ROLE_NOT_FOUND = 800,
    CANNOT_CREATE_ROLE = 801,
    CANNOT_UPDATE_ROLE = 802,
    CANNOT_DELETE_ROLE = 803,
    CANNOT_FETCH_ROLE = 804,
  
    // Password Reset Errors (9xx)
    PASSWORD_RESET_CODE_NOT_FOUND = 900,
    CANNOT_CREATE_PASSWORD_RESET_CODE = 901,
    CANNOT_FETCH_PASSWORD_RESET_CODE = 902,
    CANNOT_DELETE_PASSWORD_RESET_CODE = 903,
  
    // Email Errors (10xx)
    CANNOT_SEND_EMAIL = 1000,
    EMAIL_VERIFICATION_NOT_FOUND = 1001,
    CANNOT_CREATE_EMAIL_VERIFICATION = 1002,
    CANNOT_UPDATE_EMAIL_VERIFICATION = 1003,
    CANNOT_DELETE_EMAIL_VERIFICATION = 1004,
    CANNOT_FETCH_EMAIL_VERIFICATION = 1005,
    VERIFICATION_CODE_INVALID = 1006,
    VERIFICATION_CODE_EXPIRED = 1007,
    UNKNOWN_EMAIL_VERIFICATION_TYPE = 1008,
  
    // Employee Code Errors (11xx)
    EMPLOYEE_CODE_NOT_FOUND = 1100,
    INVALID_EMPLOYEE_CODE = 1101,
    CANNOT_CREATE_EMPLOYEE_CODE = 1102,
    CANNOT_UPDATE_EMPLOYEE_CODE = 1103,
    CANNOT_FETCH_EMPLOYEE_CODE = 1104,
    CANNOT_GENERATE_UNIQUE_CODE = 1105,
  
    // File/Image Errors (12xx)
    CANNOT_REMOVE_IMAGE_FROM_FILE = 1200,
    CANNOT_ADD_IMAGE_TO_FILE = 1201,
  
    // Validation Errors (13xx)
    VALIDATION_CODE_INCORRECT = 1300,
  
    // Other Errors (9xxx)
    UNABLE_TO_HASH = 9000,
    UNABLE_TO_ADD_TO_DB = 9001,
    CANNOT_DELETE_VERIFICATION_CODE = 9002,
  
    UNKNOWN = 9999,
  }
  
  export default ErrorCode
  